fix scala compatibility & testing & ci summary
