+++ core/src/test/scala/doric/syntax/MapColumnsSpec.scala
@@ -170,28 +170,6 @@ class MapColumnsSpec extends DoricTestElements with MapColumns {
     }
   }
 
-  describe("mapEntries doric function") {
-    import spark.implicits._
-
-    val df = List(
-      Map("k1" -> "v1", "k2" -> "v2"),
-      Map.empty[String, String],
-      null
-    ).toDF("col1")
-
-    it("should work as spark map_entries function") {
-      df.testColumns("col1")(
-        c => colMapString[String](c).mapEntries,
-        c => f.map_entries(f.col(c)),
-        List(
-          Some(Array(Row("k1", "v1"), Row("k2", "v2"))),
-          Some(Array.empty[Row]),
-          None
-        )
-      )
-    }
-  }
-
   describe("explode doric function (for maps)") {
     import spark.implicits._
